requirements.txt
setuptools
fastapi
pydantic
uvicorn
watchfiles
psycopg2-binary
debugpy